// prisma/generated/type-graphql/models/Credit_memos.d.ts
export declare class Credit_memos {
    credit_memo: string;
    order_number: number;
    created?: Date | null;
    bill_to_name: string;
    status?: string | null;
    refunded?: string | null;
    action?: string | null;
    id: bigint;
    cust_id: number;
    prod_id: bigint;
    customers?: string | null;
    products?: string | null;
}