// prisma/generated/type-graphql/resolvers/relations/Customers/args/CustomersCredit_memosArgs.d.ts
import { Credit_memosOrderByWithRelationInput } from "../../../inputs/Credit_memosOrderByWithRelationInput";
import { Credit_memosWhereInput } from "../../../inputs/Credit_memosWhereInput";
import { Credit_memosWhereUniqueInput } from "../../../inputs/Credit_memosWhereUniqueInput";
export declare class CustomersCredit_memosArgs {
    where?: Credit_memosWhereInput | undefined;
    orderBy?: Credit_memosOrderByWithRelationInput[] | undefined;
    cursor?: Credit_memosWhereUniqueInput | undefined;
    take?: number | undefined;
    skip?: number | undefined;
    distinct?: Array<"credit_memo" | "order_number" | "created" | "bill_to_name" | "status" | "refunded" | "action" | "id" | "cust_id" | "prod_id"> | undefined;
}